added alert for delete functıons

+++ src/components/AgeGroupItem.js
@@ -1,16 +1,29 @@
-// src/components/AgeGroupItem.js
 import React from 'react';
 import { List, Button } from 'react-native-paper';
 import { deleteAgeGroup } from '../services/firebaseService';
+import { Alert } from 'react-native'; // Import Alert from react-native
 
 const AgeGroupItem = ({ guideId, testId, ageGroup, navigation }) => {
-    const handleDeleteAgeGroup = async () => {
-        try {
-            await deleteAgeGroup(guideId, testId, ageGroup.id);
-            console.log(`${ageGroup.ageRange} yaş grubu silindi`);
-        } catch (error) {
-            console.error('Yaş grubu silme işlemi başarısız:', error);
-        }
+    const handleDeleteAgeGroup = () => {
+        Alert.alert(
+            'Yaş Grubu Sil',
+            `${ageGroup.ageRange} yaş grubunu silmek istediğinize emin misiniz?`,
+            [
+                { text: 'İptal', style: 'cancel' },
+                {
+                    text: 'Sil',
+                    style: 'destructive',
+                    onPress: async () => {
+                        try {
+                            await deleteAgeGroup(guideId, testId, ageGroup.id);
+                            console.log(`${ageGroup.ageRange} yaş grubu silindi`);
+                        } catch (error) {
+                            console.error('Yaş grubu silme işlemi başarısız:', error);
+                        }
+                    },
+                },
+            ]
+        );
     };
 
     return (
